♻️ reorganise request and tracing
