+++ packages/rum/src/domain/requestCollection.ts
@@ -9,9 +9,9 @@ import {
   XhrCompleteContext,
   XhrStartContext,
 } from '@datadog/browser-core'
-import { LifeCycle, LifeCycleEventType } from '../lifeCycle'
-import { startTracer, TraceIdentifier, Tracer } from '../tracer'
-import { isAllowedRequestUrl } from './resourceUtils'
+import { LifeCycle, LifeCycleEventType } from './lifeCycle'
+import { isAllowedRequestUrl } from './rumEventsCollection/resourceUtils'
+import { startTracer, TraceIdentifier, Tracer } from './tracing/tracer'
 
 export interface RequestStartEvent {
   requestIndex: number
