+++ packages/rum/src/domain/rumEventsCollection/matchRequestTiming.ts
@@ -1,5 +1,5 @@
 import { RumPerformanceResourceTiming } from '../../browser/performanceCollection'
-import { RequestCompleteEvent } from './requestCollection'
+import { RequestCompleteEvent } from '../requestCollection'
 
 interface Timing {
   startTime: number
