+++ packages/rum/src/domain/tracing/getDocumentTraceId.ts
